+++ sql/drop_db.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+sudo -u postgres psql << EOF
+drop database eva;
+drop database dispatcher;
+EOF